refactor: move delta state to autora.workflow (from workflow_)
