+++ src/autora/workflow/state/__init__.py
@@ -1,3 +1,4 @@
+from .delta import Delta, State, extend
 from .history import History
 from .param import resolve_state_params
 from .snapshot import Snapshot
