+++ src/autora/workflow_/__init__.py
@@ -1,7 +1,7 @@
+from ..workflow.state.delta import wrap_to_use_state
 from .executor import (
     experiment_runner_from_x_to_xy_function,
     experiment_runner_from_x_to_y_function,
     theorist_from_estimator,
-    wrap_to_use_state,
 )
 from .state import Delta, State
