--- src/autora/workflow_/state/__init__.py
@@ -1,1 +1,0 @@
-from .delta import Delta, Result, State, extend
